ModulartTimeSeries points()
fixes to how points are stitched together across gaps in the buffer.
Addresses edge cases where points is called with starttime=endtime, and
a single point in the buffer was not returned. Also modifies some of
the logic in the stitched points code. Seems the intent was to use this
opportunity to fill these holes in the buffer, and the changes here
assume that intent.