silly workaround for silly recursion
this commit is a little embarrassing. TimeSeries::point has to return a
0-time point for “not found” or else a clock-ed filter will recurse
past the end of available data. But this necessitates some other
changes to some of the other ts filters, namely, the outlier excluder
and the valid ranger. clock-ed filters that can optionally drop points
are problematic because we never know if they will need to expand their
source query range, or by how far, before the actual exclusion is done.
the valid ranger is simpler since it’s a single point filter. the
outlier excluder is less simple, because what if everything is an
outlier? in this case, we’re just expanding the search window by 3x and
if there’s not enough data to resample, then we sort of give up. maybe
make this smarter in the future?