Fixes stepSimulation() error "Simulation step used ... different than expected"
I don’t think we need to set the report step here, in which case I
believe it will always stay at the hydraulic step, which should be
fine. Setting it here results in Rstep being set to < Hstep values
during short time steps. At the next step this can lead to the tank
updates being restricted by this short Rstep, and when that is
controlling you get the above error.

In short, this is all due to interaction with the great Epanet hairball.