Don't try to set simulation parameters for valve status
The current Epanet methods for setting a valve simple control  can not discern between changing of a status (Closed / Open / Active) and a setting. This is because the API is written to only accept one parameter.

Definitely fixable, but annoyingly complicated, and again lipstick on a pig. We can get by no problem without being able to close/open valves.