fixes handling of valve/pipe setting/status changes
there's still a little weirdness here, since RTX conflicts with how EPANET wants to comprehend setting/status states. Internally EPANET will track status not as just OPEN/CLOSED, but also ACTIVE and a whole lot of other things. So for now, when we change a status to CLOSED, the setting should not be touched. And when status is "normal" or open/active, then the setting will be assigned.